+++ app/src/main/res/drawable/icon_back.xml
@@ -1,0 +1,5 @@
+<vector android:height="24dp" android:tint="#706C6C"
+    android:viewportHeight="24" android:viewportWidth="24"
+    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="@android:color/white" android:pathData="M17.77,3.77l-1.77,-1.77l-10,10l10,10l1.77,-1.77l-8.23,-8.23z"/>
+</vector>
