+++ app/src/main/res/layout/book_item_layout.xml
@@ -14,7 +14,7 @@
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:background="@color/white">
+            >
 
             <ImageView
                 android:id="@+id/iv_icon"
@@ -30,10 +30,10 @@
                 android:id="@+id/tv_name"
                 android:layout_width="200dp"
                 android:layout_height="wrap_content"
-                android:layout_marginStart="5dp"
+                android:layout_marginStart="15dp"
                 android:gravity="start"
+                style="@style/content_regular"
                 app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toEndOf="@id/iv_icon"
                 app:layout_constraintTop_toTopOf="parent"
