+++ app/src/main/res/values-night/themes.xml
@@ -10,9 +10,25 @@
         <!-- Secondary brand color. -->
         <item name="colorSecondary">@color/teal_200</item>
         <item name="colorSecondaryVariant">@color/teal_200</item>
-        <item name="colorOnSecondary">@color/black</item>
+        <item name="colorOnSecondary">#F8F6F6</item>
         <!-- Status bar color. -->
         <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
         <!-- Customize your theme here. -->
+    </style>
+
+    <style name="custom_card_view" parent="CardView">
+        <item name="backgroundColor">@color/gray</item>
+        <item name="android:layout_marginTop">@dimen/general_vertical_margin_min</item>
+        <item name="android:layout_marginBottom">@dimen/general_vertical_margin</item>
+        <item name="android:layout_marginStart">@dimen/general_horizontal_margin</item>
+        <item name="android:layout_marginEnd">@dimen/general_horizontal_margin</item>
+        <item name="android:padding">5dp</item>
+        <item name="cardCornerRadius">@dimen/general_corner_radius</item>
+    </style>
+
+    <style name="sub_title_min" parent="Theme.BazAndroidCourse">
+        <item name="android:textSize">@dimen/font_content_footer</item>
+        <item name="marginHorizontal">@dimen/general_horizontal_margin</item>
+        <item name="android:textColor">@color/green</item>
     </style>
 </resources>
